feat: add talent messages page and role-based routing
