+++ talentify-next-frontend/app/talent/messages/page.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import MessagesPage from '@/components/messages/MessagesPage'
+
+export default function TalentMessagesPage() {
+  return <MessagesPage role="talent" />
+}
+
