+++ talentify-next-frontend/app/talent/notifications/page.tsx
@@ -47,7 +47,7 @@ export default function TalentNotificationsPage() {
   }
 
   const linkFor = (n: NotificationRow) => {
-    if (n.type === 'message') return '/messages'
+    if (n.type === 'message') return '/talent/messages'
     const offerId = (n.data as any)?.offer_id
     return offerId ? `/talent/offers/${offerId}` : '#'
   }
