+++ talentify-next-frontend/utils/messages.ts
@@ -4,13 +4,13 @@ import { createClient } from '@/utils/supabase/client'
 
 const supabase = createClient()
 
-export async function getUnreadMessageCount(): Promise<number> {
+export async function getUnreadMessageCount(role: 'store' | 'talent'): Promise<number> {
   const {
     data: { user },
   } = await supabase.auth.getUser()
   if (!user) return 0
 
-  const { data, error } = await supabase.rpc('unread_messages_count')
+  const { data, error } = await supabase.rpc('unread_messages_count', { role })
   if (error) {
     console.error('failed to fetch unread messages count', error)
     return 0
